+++ app/user/settings/page.js
@@ -3,7 +3,9 @@
 import React, { useState } from 'react';
 import { Box, Button, Typography, Paper, FormControl, InputLabel, MenuItem, Select, Stack } from '@mui/material';
 import { useRouter } from 'next/navigation';
-import SidebarUser from '../../components/SidebarUser';
+import StyledPaper from "../../components/styledComponents/StyledPaper";
+import {BlueButton,GreenButton ,RedButton} from "../../components/styledComponents/StyledButton";
+import DesignTitel from "../../components/styledComponents/DesignTitel";
 
 function UserSettings() {
   const router = useRouter();
@@ -27,12 +29,12 @@ function UserSettings() {
   };
 
   return (
-    <Box>
+    <StyledPaper>
 
       <Box>
-        <Typography variant="h4" gutterBottom sx={{ marginBottom: 4, textAlign: 'center', fontWeight: 'bold', color: '#333' }}>
+        <DesignTitel>
           Einstellungen
-        </Typography>
+        </DesignTitel>
 
         <Paper
           elevation={4}
@@ -75,26 +77,20 @@ function UserSettings() {
           </FormControl>
 
           <Stack direction="row" justifyContent="space-between" spacing={2} mt={2}>
-            <Button
+            <RedButton
               onClick={handleGoBackClick}
-              variant="outlined"
-              color="error"
-              sx={{ fontWeight: 'bold', textTransform: 'none' }}
             >
               Zurück
-            </Button>
-            <Button
+            </RedButton>
+            <GreenButton
               onClick={handleSaveChangesClick}
-              variant="contained"
-              color="success"
-              sx={{ fontWeight: 'bold', textTransform: 'none' }}
             >
               Änderungen speichern
-            </Button>
+            </GreenButton>
           </Stack>
         </Paper>
       </Box>
-    </Box>
+    </StyledPaper>
   );
 }
